# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Unisalones API",
    "version": "1.0.0",
    "description": "API REST del Sistema de Gestión de Espacios - Unicomfacauca (Unisalones)",
    "contact": {
      "name": "Equipo Unisalones"
    }
  },
  "servers": [
    {
      "url": "http://localhost:{port}/api",
      "variables": {
        "port": {
          "default": "3000"
        }
      },
      "description": "Servidor local (desarrollo)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Introduce el token con el formato: Bearer <tu_token_JWT>"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/spaces/available": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "summary": "Consultar espacios disponibles",
        "description": "Devuelve la disponibilidad de los espacios según los filtros de fecha u hora.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Fecha para verificar la disponibilidad (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "start",
            "schema": {
              "type": "string",
              "example": "08:00"
            },
            "description": "Hora inicial del rango de búsqueda (HH:mm)"
          },
          {
            "in": "query",
            "name": "end",
            "schema": {
              "type": "string",
              "example": "18:00"
            },
            "description": "Hora final del rango de búsqueda (HH:mm)"
          },
          {
            "in": "query",
            "name": "spaceId",
            "schema": {
              "type": "integer"
            },
            "description": "ID del espacio específico a consultar"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de espacios con sus intervalos disponibles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "spaceId": {
                        "type": "integer"
                      },
                      "slots": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "start": {
                              "type": "string"
                            },
                            "end": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string",
                              "enum": [
                                "available",
                                "reserved"
                              ]
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Petición inválida"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/spaces": {
      "post": {
        "tags": [
          "Spaces"
        ],
        "summary": "Crear un nuevo espacio",
        "description": "Solo los administradores pueden crear nuevos espacios.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "capacity"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Espacio creado correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "Prohibido"
          }
        }
      },
      "get": {
        "tags": [
          "Spaces"
        ],
        "summary": "Listar todos los espacios",
        "description": "Devuelve todos los espacios registrados.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de espacios"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/spaces/{id}": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "summary": "Obtener un espacio por ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del espacio"
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del espacio encontrado"
          },
          "404": {
            "description": "Espacio no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Spaces"
        ],
        "summary": "Actualizar un espacio",
        "description": "Solo administradores pueden modificar los datos de un espacio.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del espacio a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Espacio actualizado correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "403": {
            "description": "No autorizado"
          }
        }
      },
      "delete": {
        "tags": [
          "Spaces"
        ],
        "summary": "Eliminar un espacio",
        "description": "Solo administradores pueden eliminar espacios.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del espacio a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Espacio eliminado correctamente"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "Prohibido"
          }
        }
      }
    },
...
  }
}